refactor to create UserService
