+++ webapp/src/app/app.component.html
@@ -2,17 +2,17 @@
   {{title}}
 </h1>
 
-<div *ngIf="loginStatus == 'Unknown'">
+<div *ngIf="userService.loginStatus == 'Unknown'">
   <p>loading...</p>
 </div>
 
-<div *ngIf="loginStatus == 'LoggedOut'">
+<div *ngIf="userService.loginStatus == 'LoggedOut'">
   <button (click)="login()" md-button>Log In</button>
 </div>
 
-<div *ngIf="loginStatus == 'LoggedIn'">
-  <div *ngIf="(profile$ | async)">
-    Profile Info: {{ (profile$ | async) | json }}
+<div *ngIf="userService.loginStatus == 'LoggedIn'">
+  <div *ngIf="(userService.profile$ | async)">
+    Profile Info: {{ (userService.profile$ | async) | json }}
   </div>
   <button (click)="logout()" md-button>Log Out</button>
 </div>
